requirements.txt
numpy
pandas
scikit-learn
jupyterlab
streamlit
requests
python-dotenv